refactor: provider to manager mixin
